Add files via upload
This Analysis found the result that #Smallcap stocks had higher risk-adjusted returns than #Largecap stocks all the way back from 1926.

1. Understanding the dataset.
This is a very big dataset and it goes all the way down to 1926. So there is a row for every month, 1926,07,08,09, all the way through December of 2018. The first column is the date. Other columns are a bunch of portfolios. It divides the stocks into three categories by market cap. The lowest 30 percent, the middle 40 percent, and the highest 30 percent. And then taken all the highest 30 percent stocks, weighted them equally, and build a portfolio, and that is the monthly return of that portfolio. The next set of returns here is quintiles. They are broken into 5 groups, the lowest 20 percent, Quintile 2, Quintile 3, Quintile 4, and the high 20 percent. And the next one is of deciles.

2. Imported the data using pd.read_csv.
3. Analysed two datasets,
(a). Low 10% market cap stocks portfolio and high 10% market cap stocks portfolio.
(b). Low 20% market cap stocks portfolio and high 20% market cap stocks portfolio.
3. Calculated Annualized Returns and Annualized Volatility of the Smallcap and Largecap Portfolios for the period of 1926-2018.
4. Made the plot.
5. Calculated risk-adjusted returns and found that small-cap portfolios delivered high risk-adjusted returns than large-cap portfolios for both datasets over the analyzed time period.